// PC.BE/src/common/enums/target.enum.ts
export enum TargetSliders {
  bannerHome = 'banner::home',
  bannerProfits = 'banner::profits',
  bannerPromotions = 'banner::promotions',
  bannerFinancially = 'banner::financially',
  bannerBusiness = 'banner::business',
  bannerEnterprise = 'banner::enterprise',
  bannerInsurance = 'banner::insurance',
  bannerProuser = 'banner::prouser',
  bannerAdjudicated = 'banner::adjudicated',
}

// TARGET PARENT FOR CATEGORY
export enum TargetParentCategories {
  // PRODUCTS PARENTS CATEGORIES
  categoryGlobal = 'category::global',
  categoryCards = 'category::cards',
  categoryAccounts = 'category::accounts',
  categoryLoans = 'category::loans',
  categoryDeposits = 'category::deposits',
  categoryPointsBSC = 'category::points-bsc',
}

// TARGET FOR CATEGORY
export enum TargetCategories {
  categoryProfits = 'category::profits',
  categoryPodcast = 'category::podcast',
  categoryProducts = 'category::products',
  categoryAboutUs = 'category::about-us',
  categoryBusiness = 'category::business',
  categoryEnterprise = 'category::enterprise',
  categoryChannels = 'category::channels',
  categoryInsurance = 'category::insurance',
  categoryProuser = 'category::prouser',
  categoryAdjudicated = 'category::adjudicated',
}

// FEATURED GLOBALS POSITIONS
export enum GlobalPositionsFreatured {
  hidden = 'hidden',
  menu = 'menu',
  menuExpanded = 'menu-expanded',
  footer = 'footer',
  sidebar = 'sidebar',
  mainContent = 'main-content',
  recommended = 'recommended',
}